Updated project to Vite version

+++ OneDrive/Desktop/student-learning/src/main/java/com/example/demo/repository/SavedVideoRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.SavedVideo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SavedVideoRepository
-        extends JpaRepository<SavedVideo, Integer>{}
+public interface SavedVideoRepository extends JpaRepository<SavedVideo, Long> {
+}
